added (back) files to run make cat with sm=0

+++ scripts/sh/init_run_exclusive_canfar.sh
@@ -13,6 +13,7 @@ debug_out=-1
 scratch=-1
 fix=0
 test_only=0
+sm=1
 
 # mh_local is 0 (1) if merge_header_runner is run on all exposures,
 # which is standard so far (run on exposures of given tile only; new)
@@ -39,6 +40,8 @@ usage="Usage: $(basename "$0") -j JOB -e ID -k KIND [OPTIONS]
    \tmerge header file local (MH=1) or global (MH=0); default is $mh_local\n
    -s, --sp_local SP\n
    \tsplit local run local (SP=1) or global (SP=r0wwdefault is $sp_local\n
+   --sm SM\n
+   \tWith (SM=1; default) or without (SM=0) spread model input\n
    -N, --N_SMP N_SMOp\n
     \tnumber of jobs (SMP mode only), default from original config files\n
    -d, --directory\n
@@ -86,6 +89,10 @@ while [ $# -gt 0 ]; do
       ;;
     -s|--sp_local)
       sp_local="$2"
+      shift
+      ;;
+    --sm)
+      sm="$2"
       shift
       ;;
     -N|--N_SMP)                                                                 
@@ -139,6 +146,7 @@ function message() {
 
 
 # Init message
+message "test=$test_only" $debug_out -1
 if [ "$test_only" == "1" ]; then
   msg="init_run_exclusive.py script test mode, exiting."
   ex=0
@@ -202,19 +210,10 @@ else
   message "not running in dry run mode" $debug_out -1
 fi
 
-source activate shapepipe
-if [ "$debug_out"  != "-1" ]; then
-    echo "${pat}conda prefix = ${CONDA_PREFIX}" >> $debug_out
-fi
-
 CONDA_PREFIX=$HOME/.conda/envs/shapepipe
 PATH=$PATH:$CONDA_PREFIX/bin
-message "conda prefix = ${CONDA_PREFIX}" $debug_out -1
-message "HOME = ${HOME}" $debug_out -1
-message "path = ${PATH}" $debug_out -1
 
 cd $dir
-message "pwd=$pwd" $debug_out -1
 
 if [ ! -d ${kind}_runs ]; then
   command "mkdir ${kind}_runs" $dry_run
@@ -358,7 +357,7 @@ if [ $do_job != 0 ] && [ "$sp_local" == "1" ]; then
   # run local Sp if not done already; works only with mh_local=1; this step needs to be done
   # before following mh_local=1 steps 
   message "run local sp" $debug_out -1
-  command "rm -rf run_sp_GitFeGie*/get_images_runner_run_2" $dry_run
+  #command "rm -rf run_sp_GitFeGie*/get_images_runner_run_2" $dry_run
   command "rm -rf run_sp_Gie*" $dry_run
   command "rm -rf run_sp_exp_Sp*" $dry_run
 
@@ -411,10 +410,6 @@ if [ $do_job != 0 ] && [ "$sp_local" == "1" ]; then
 fi
 
 if [ "$kind" == "tile" ] && [ "$sp_local" == "1" ]; then
-  # Link to exposure outputs (runs 2 and 32)
-
-  # Remove previous split exp dir
-rm -r run_sp_exp_Sp_shdu
   cd ../../..
   command "link_to_exp_for_tile.py -t $ID -i tile_runs -I exp_runs -s $sp_local" $dry_run
   cd tile_runs/$ID
@@ -439,6 +434,7 @@ else
     message "ID needs to be given (option -e) for mh_local" $debug_out 6
   fi
 
+  # Check and remove symbolic (global) mh file link
   if [ -L log_exp_headers.sqlite ]; then
     # Local Mh and symlink -> remove previous link to
     # (potentially incomplete) global mh file
@@ -446,6 +442,17 @@ else
     command "rm log_exp_headers.sqlite" $dry_run
   else
     message "no mh link found" $debug_out -1
+  fi
+
+  # Check size of existing header file
+  if [ -e log_exp_headers.sqlite ]; then
+    size=$(stat -c %s log_exp_headers.sqlite)
+    if (( size > 15000 )); then
+      message "Found valid local mh file, continuing" $debug_out -1
+    else
+      message "Existing local mh file looks invalid, deleting" $debug_out -1
+      rm -f log_exp_headers.sqlite
+    fi
   fi
 
   if [ ! -e log_exp_headers.sqlite ]; then
@@ -478,58 +485,23 @@ fi
 # Update links to exposure run directories, which were created in job 32
 (( do_job = $job & 64 ))
 if [[ $do_job != 0 ]]; then
+  if [ "$kind" == "tile" ]; then
+    cd ../../..
+    command "link_to_exp_for_tile.py -t $ID -i tile_runs -I exp_runs -s $sp_local" $dry_run
+    cd ${kind}_runs/$ID/output
 
-  # MKDEBUG NEW (04/01/2025): this script was already run earlier with -s 1
-  #cd ../../..
-  #command "link_to_exp_for_tile.py -t $ID -i tile_runs -I exp_runs -s $sp_local" $dry_run
-  #cd ${kind}_runs/$ID/output
+    # Remove duplicate job-16 runs (tile detection)
+    # New (P8) commented
+    #n_16=`ls -rt1d run_sp_tile_Sx_* | wc -l`
+    #if [ "$n_16" != "1" ]; then
+      #n_remove="$(($n_16-1))"
+      #echo "removing $n_remove duplicate old job-16 runs"
+      #command "rm -rf `ls -rt1d run_sp_tile_Sx_* | head -$n_remove`" $dry_run
+    #fi
 
-  # Remove previous runs of this job
-  rm -rf run_sp_tile_PsViSmVi*
-fi
-
-(( do_job = $job & 128 ))
-if [[ $do_job != 0 ]]; then
-
-    echo
-    
-    cat_ngmix="run_sp_tile_ngmix_Ng1u/ngmix_runner/output/ngmix-*.fits"
-    dir_ngmix_prev="run_sp_tile_ngmix_Ng1u_prev/ngmix_runner/output"
-    cat_ngmix_prev="$dir_ngmix_prev/ngmix-*.fits"
-    
-    # Check whether ngmix output exists
-    if [ -e $cat_ngmix ]; then
-        message "ngmix output catalogue exists" $debug_out -1
-
-        # Check whether previous ngmix directory and output cat exist
-        exists="1"
-        if [ ! -d $dir_ngmix_prev ]; then
-            exists="0"
-        elif [ ! -e "$cat_ngmix_prev" ]; then
-            exists="1"
-        fi
-        if [ "$exists" == "0" ]; then
-            message "Moving to previous batch-save dir (does not exist yet)" $debug_out -1
-            command "mkdir -p $dir_ngmix_prev" $dry_run
-            command "mv $cat_ngmix $dir_ngmix_prev" $dry_run
-        else
-            # Compare file sizes
-            size_cat_ngmix=$(stat -c%s $cat_ngmix)
-            size_cat_ngmix_prev=$(stat -c%s $cat_ngmix_prev)
-            if [ "$size_cat_ngmix" -gt "$size_cat_ngmix_prev" ]; then
-                message "Moving to batch-save dir, overwriting smaller batch-save cat" $debug_out -1
-                command "mv $cat_ngmix $dir_ngmix_prev" $dry_run
-            else
-                message "Previous batch-save dir not smaller, removing ngmix output" $debug_out -1
-                command "rm $cat_ngmix" $dry_run
-            fi
-        fi
-    else
-        # Whether or not previous ngmix exists, job_sp_canfar will handle it
-        message "No ngmix output exists, continuing..." $debug_out -1
-    fi
-
-    echo
+    # Remove previous runs of this job
+    rm -rf run_sp_tile_PsViSmVi*
+  fi
 fi
 
 (( do_job = $job & 256 ))
@@ -537,6 +509,7 @@ if [[ $do_job != 0 ]]; then
 
   # Remove previous runs of this job
   rm -rf run_sp_Ms_20??_*
+
 fi
 
 (( do_job = $job & 512 ))
@@ -572,11 +545,13 @@ if [ "$scratch" != "-1" ]; then
   command "cd $scratch/${kind}_runs/$ID" $dry_run
 fi
 
-command "job_sp_canfar.bash -p psfex -j $job -e $ID --n_smp $N_SMP --nsh_jobs $N_SMP --debug_out $debug_out " $dry_run
+command "job_sp_canfar.bash -p psfex -j $job -e $ID --n_smp $N_SMP --nsh_jobs $N_SMP --debug_out $debug_out --sm $sm " $dry_run
 
 if [ "$scratch" != "-1" ]; then
   cd ../..
-  if [ "$job" == "32" ]; then                                                   
+  if [ "$job" == "16" ]; then                                                   
+    command "mv ${kind}_runs/$ID/output/run_sp_Sx_* $dir/${kind}_runs/$ID/output" $dry_run
+  elif [ "$job" == "32" ]; then                                                   
     command "mv ${kind}_runs/$ID/output/run_sp_exp_SxSe* $dir/${kind}_runs/$ID/output" $dry_run
   elif [ "$job" == "64" ]; then                                                 
     command "mv ${kind}_runs/$ID/output/run_sp_tile_PsViSm** $dir/${kind}_runs/$ID/output" $dry_run
